+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## 2.5.0
 
-* Added new DecimalFields for both public and symmetric key
-* Added new FloatFields for both public and symmetric key
-* Added new TimeFields for both public and symmetric key
+* Added new DecimalFields for both public and symmetric key (#64)
+* Added new FloatFields for both public and symmetric key (#64)
+* Added new TimeFields for both public and symmetric key (#64)
 
 ## 2.4.0